Moved `Types` & `Interface` into a different file
